Collection names instead of IDs in search results
Look up AccessSiteCollection Name from the db based on search results collection id.

It's imperfect to use icontains as it could math substrings, but
considering the small number of collection, probably not a concern.

Also, switch date rendering to a templatetag for more safety

Use the `/fr/search` endpoint for searching when document language is `fr`